+++ packages/python_ffi_cpython/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Refactored to make use of Dart 3.x record types.
 - Fixed path issues on Windows.
+- Fixed
+  [Bundling Python modules with external package](https://github.com/IVLIVS-III/dart_python_ffi/issues/10)
+  .
 
 ## 0.3.1