+++ packages/python_ffi_interface/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Refactored to make use of Dart 3.x record types.
 - Utilized Python builtin function `iter` to convert Python iterables to Dart iterables.
+- Fixed
+  [Bundling Python modules with external package](https://github.com/IVLIVS-III/dart_python_ffi/issues/10)
+  .
 
 ## 0.1.1